requirements.txt
pymysql==1.1.0
pillow==10.2.0
pandas==2.2.0
matplotlib==3.8.2
tkcalendar==1.6.1
schedule==1.2.1